+++ ui/formatter_test.go
@@ -1,4 +1,4 @@
-package ioutil
+package ui
 
 import (
 	"testing"
